Update receipt result class
Translate comments and add fields.
Add raw content field.
