+++ CheckReceiptSDK/Results/ReceiptResult.cs
@@ -5,14 +5,18 @@ using System.Runtime.Serialization;
 namespace CheckReceiptSDK.Results
 {
     /// <summary>
-    /// Класс, описывающий информацию, получаемую в результате запроса от ФНС детальной информации по чеку
+    /// Represents information received from Federal Tax Service when requested detailed information.
     /// </summary>
     public sealed class ReceiptResult : Result
     {
         /// <summary>
-        /// Информация о документе, которая приходит из ФНС
+        /// Document information received from FTS
         /// </summary>
         public Document Document { get; internal set; }
+        /// <summary>
+        /// Raw response
+        /// </summary>
+        public string Content { get; internal set; }
 
         internal ReceiptResult()
         { }
@@ -25,40 +29,41 @@ namespace CheckReceiptSDK.Results
         public Document Document { get; internal set; }
     }
     /// <summary>
-    /// Документ, который приходит из ФНС
+    /// Document information received from FTS
     /// </summary>
     [DataContract]
     public class Document
     {
         /// <summary>
-        /// Внутренняя информация о чеке
+        /// Receipt's information
         /// </summary>
         [DataMember]
         public Receipt Receipt { get; internal set; }
     }
     /// <summary>
-    /// Непосредственно сам чек. В разных чеках по разному заполнены параметры.
+    /// Represents receipt. IN different receipts parameters can be filled differently.
     /// </summary>
     [DataContract]
     public class Receipt
     {
         #region Money
         /// <summary>
-        /// Общая сумма по чеку, в копейках
+        /// Total amount in kopecks
         /// </summary>
         [DataMember]
         public int TotalSum { get; internal set; }
         /// <summary>
-        /// Сумма, оплаченная наличными, в копейках
+        /// Amount paid by cash in kopecks
         /// </summary>
         [DataMember]
         public int CashTotalSum { get; internal set; }
         /// <summary>
-        /// Сумма, оплаченная безналичным способом оплаты, в копейках
+        /// Amount paid by card in kopecks
         /// </summary>
         [DataMember]
         public int EcashTotalSum { get; internal set; }
         /// <summary>
+        /// VAT
         /// Сумма НДС оплаченная по ставке 18%, в копейках
         /// </summary>
         [DataMember(Name = "nds18")]
@@ -186,53 +191,5 @@ namespace CheckReceiptSDK.Results
         [DataMember(IsRequired = false)]
         public List<object> Properties { get; internal set; }
         #endregion
-    }
-
-    /// <summary>
-    /// Класс, представляющий позицию в чеке. В разных чеках параметры заполнены по разному.
-    /// </summary>
-    [DataContract]
-    public class Item
-    {
-        /// <summary>
-        /// Сумма по позиции, в копейках
-        /// </summary>
-        [DataMember]
-        public int Sum { get; internal set; }
-        /// <summary>
-        /// Количество
-        /// </summary>
-        [DataMember]
-        public double Quantity { get; internal set; }
-        /// <summary>
-        /// Цена позиции, в копейках
-        /// </summary>
-        [DataMember]
-        public int Price { get; internal set; }
-        /// <summary>
-        /// Наименование позиции
-        /// </summary>
-        [DataMember]
-        public string Name { get; internal set; }
-        /// <summary>
-        /// Сумма НДС, оплаченная по ставке 10%, в копейках
-        /// </summary>
-        [DataMember(Name = "nds10", IsRequired = false)]
-        public int Nds10Sum { get; internal set; }
-        /// <summary>
-        /// Сумма НДС, оплаченная по ставке 18%, в копейках
-        /// </summary>
-        [DataMember(Name = "nds18", IsRequired = false)]
-        public int Nds18Sum { get; internal set; }
-        /// <summary>
-        /// Не понимаю что это
-        /// </summary>
-        [DataMember(IsRequired = false)]
-        public List<object> Properties { get; internal set; }
-        /// <summary>
-        /// Не понимаю что это
-        /// </summary>
-        [DataMember(IsRequired = false)]
-        public List<object> Modifiers { get; internal set; }
     }
 }
